pyproject.toml
[build-system]
requires = ["flit_core>=3.2", "flit_scm", "wheel"]
build-backend = "flit_scm:buildapi"

[project]
name = "django-tom-select"
authors = [
  { name = "David Rydwanski", email = "dr@pader-solartechnik.de" },
]
readme = "README.rst"
license = { file = "LICENSE" }
keywords = ["Django", "tom-select", "autocomplete", "typeahead"]
dynamic = ["version", "description"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Developers",
  "Environment :: Web Environment",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Framework :: Django",
  "Framework :: Django :: 4.2",
  "Topic :: Software Development",
]
requires-python = ">=3.8"
dependencies = [
  "django>=4.2",
  "django-appconf>=0.6.0"
]

[project.optional-dependencies]
test = [
  "pytest",
  "pytest-cov",
  "pytest-django",
  "selenium",
]
selenium = [
  "selenium",
]
docs = [
  "sphinx",
]

[project.urls]
Project-URL = "https://github.com/codingjoe/django-tom-select"
Changelog = "https://github.com/codingjoe/django-tom-select/releases"
Source = "https://github.com/codingjoe/django-tom-select"
Documentation = "https://django-tom-select.rtfd.io/"
Issue-Tracker = "https://github.com/codingjoe/django-tom-select/issues"

[tool.flit.module]
name = "django_tom_select"

[tool.setuptools_scm]
write_to = "django_tom_select/_version.py"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov --tb=short -rxs"
testpaths = ["tests"]
DJANGO_SETTINGS_MODULE = "tests.testapp.settings"
filterwarnings = ["ignore::PendingDeprecationWarning", "error::RuntimeWarning"]

[tool.coverage.run]
source = ["django_tom_select"]

[tool.coverage.report]
show_missing = true

[tool.isort]
atomic = true
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_first_party = "django_tom_select, tests"
default_section = "THIRDPARTY"
combine_as_imports = true

[tool.pydocstyle]
add_ignore = "D1"
